working after refactor
